backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic-settings==2.13.1
pytest==8.4.2
httpx==0.28.1
SQLAlchemy==2.0.48
alembic==1.18.4
psycopg[binary]==3.3.3
celery==5.6.2
redis==7.3.0
trafilatura==2.0.0
beautifulsoup4==4.14.3
openai==2.26.0
